refactor: Sections 일급 컬렉션 추가
