+++ src/main/java/nextstep/subway/line/domain/Line.java
@@ -14,9 +14,8 @@ public class Line extends BaseEntity {
     @Column(unique = true)
     private String name;
     private String color;
-
-    @OneToMany(mappedBy = "line", cascade = {CascadeType.PERSIST, CascadeType.MERGE}, orphanRemoval = true)
-    private List<Section> sections = new ArrayList<>();
+    @Embedded
+    private Sections sections = new Sections();
 
     public Line() {
     }
@@ -29,7 +28,7 @@ public class Line extends BaseEntity {
     public Line(String name, String color, Station upStation, Station downStation, int distance) {
         this.name = name;
         this.color = color;
-        sections.add(new Section(this, upStation, downStation, distance));
+        sections.add(this, upStation, downStation, distance);
     }
 
     public void update(Line line) {
@@ -49,128 +48,15 @@ public class Line extends BaseEntity {
         return color;
     }
 
-    public List<Section> getSections() {
-        return sections;
-    }
-
     public List<Station> getStations() {
-        if (sections.isEmpty()) {
-            return Collections.emptyList();
-        }
-        return getStationsInOrder();
-    }
-
-    private List<Station> getStationsInOrder() {
-        List<Station> stations = new ArrayList<>();
-        Station downStation = findUpStation();
-        stations.add(downStation);
-
-        while (downStation != null) {
-            Station finalDownStation = downStation;
-            Optional<Section> nextLineStation = sections.stream()
-                    .filter(it -> it.getUpStation() == finalDownStation)
-                    .findFirst();
-            if (!nextLineStation.isPresent()) {
-                break;
-            }
-            downStation = nextLineStation.get().getDownStation();
-            stations.add(downStation);
-        }
-
-        return stations;
-    }
-
-    public Station findUpStation() {
-        Station downStation = sections.get(0).getUpStation();
-        while (downStation != null) {
-            Station finalDownStation = downStation;
-            Optional<Section> nextLineStation = sections.stream()
-                    .filter(it -> it.getDownStation() == finalDownStation)
-                    .findFirst();
-            if (!nextLineStation.isPresent()) {
-                break;
-            }
-            downStation = nextLineStation.get().getUpStation();
-        }
-
-        return downStation;
+        return sections.getStations();
     }
 
     public void addSection(Station upStation, Station downStation, int distance) {
-        boolean isUpStationExisted = isExisted(upStation);
-        boolean isDownStationExisted = isExisted(downStation);
-
-        validate(isUpStationExisted, isDownStationExisted);
-
-        if (isUpStationExisted) {
-            updateUpStation(upStation, downStation, distance);
-        }
-        if (isDownStationExisted) {
-            updateDownStation(upStation, downStation, distance);
-        }
-
-        sections.add(new Section(this, upStation, downStation, distance));
-    }
-
-    private void validate(boolean isUpStationExisted, boolean isDownStationExisted) {
-        if (isUpStationExisted && isDownStationExisted) {
-            throw new RuntimeException("이미 등록된 구간 입니다.");
-        }
-        if (!getStations().isEmpty() && !isUpStationExisted && !isDownStationExisted) {
-            throw new RuntimeException("등록할 수 없는 구간 입니다.");
-        }
-    }
-
-    private void updateDownStation(Station upStation, Station downStation, int distance) {
-        sections.stream()
-                .filter(it -> it.getDownStation() == downStation)
-                .findFirst()
-                .ifPresent(it -> it.updateDownStation(upStation, distance));
-    }
-
-    private void updateUpStation(Station upStation, Station downStation, int distance) {
-        sections.stream()
-                .filter(it -> it.getUpStation() == upStation)
-                .findFirst()
-                .ifPresent(it -> it.updateUpStation(downStation, distance));
-    }
-
-    private boolean isExisted(Station station) {
-        return getStations().stream().anyMatch(it -> it == station);
+        sections.add(this, upStation, downStation, distance);
     }
 
     public void removeLineStation(Long stationId) {
-        if (sections.size() <= 1) {
-            throw new IllegalArgumentException("역을 제거할 수 없습니다.");
-        }
-
-        Optional<Section> upLineStation = findUpStation(stationId);
-        Optional<Section> downLineStation = findDownStation(stationId);
-
-        if (upLineStation.isPresent() && downLineStation.isPresent()) {
-            createSection(upLineStation.get(), downLineStation.get());
-        }
-
-        upLineStation.ifPresent(it -> sections.remove(it));
-        downLineStation.ifPresent(it -> sections.remove(it));
-    }
-
-    private void createSection(Section upLineStation, Section downLineStation) {
-        Station newUpStation = downLineStation.getUpStation();
-        Station newDownStation = upLineStation.getDownStation();
-        int newDistance = upLineStation.getDistance() + downLineStation.getDistance();
-        sections.add(new Section(this, newUpStation, newDownStation, newDistance));
-    }
-
-    private Optional<Section> findDownStation(Long stationId) {
-        return sections.stream()
-                .filter(it -> it.equalsDownStation(stationId))
-                .findFirst();
-    }
-
-    private Optional<Section> findUpStation(Long stationId) {
-        return sections.stream()
-                .filter(it -> it.equalsUpStation(stationId))
-                .findFirst();
+        sections.removeLineStation(this, stationId);
     }
 }
