+++ src/main/java/nextstep/subway/line/domain/Sections.java
@@ -1,0 +1,138 @@
+package nextstep.subway.line.domain;
+
+import nextstep.subway.station.domain.Station;
+
+import javax.persistence.CascadeType;
+import javax.persistence.Embeddable;
+import javax.persistence.OneToMany;
+import java.util.ArrayList;
+import java.util.Collections;
+import java.util.List;
+import java.util.Optional;
+
+@Embeddable
+public class Sections {
+    @OneToMany(mappedBy = "line", cascade = {CascadeType.PERSIST, CascadeType.MERGE}, orphanRemoval = true)
+    private List<Section> sections = new ArrayList<>();
+
+    public void add(Line line, Station upStation, Station downStation, int distance) {
+        boolean isUpStationExisted = isExisted(upStation);
+        boolean isDownStationExisted = isExisted(downStation);
+
+        validate(isUpStationExisted, isDownStationExisted);
+
+        if (isUpStationExisted) {
+            updateUpStation(upStation, downStation, distance);
+        }
+        if (isDownStationExisted) {
+            updateDownStation(upStation, downStation, distance);
+        }
+
+        sections.add(new Section(line, upStation, downStation, distance));
+    }
+
+    public List<Station> getStations() {
+        if (sections.isEmpty()) {
+            return Collections.emptyList();
+        }
+        return getStationsInOrder();
+    }
+
+    public void removeLineStation(Line line, Long stationId) {
+        if (sections.size() <= 1) {
+            throw new IllegalArgumentException("역을 제거할 수 없습니다.");
+        }
+
+        Optional<Section> upLineStation = findUpStation(stationId);
+        Optional<Section> downLineStation = findDownStation(stationId);
+
+        if (upLineStation.isPresent() && downLineStation.isPresent()) {
+            createSection(line, upLineStation.get(), downLineStation.get());
+        }
+
+        upLineStation.ifPresent(it -> sections.remove(it));
+        downLineStation.ifPresent(it -> sections.remove(it));
+    }
+
+    private Station findUpStation() {
+        Station downStation = sections.get(0).getUpStation();
+        while (downStation != null) {
+            Station finalDownStation = downStation;
+            Optional<Section> nextLineStation = sections.stream()
+                    .filter(it -> it.getDownStation() == finalDownStation)
+                    .findFirst();
+            if (!nextLineStation.isPresent()) {
+                break;
+            }
+            downStation = nextLineStation.get().getUpStation();
+        }
+
+        return downStation;
+    }
+
+    private boolean isExisted(Station station) {
+        return getStations().stream().anyMatch(it -> it == station);
+    }
+
+    private void validate(boolean isUpStationExisted, boolean isDownStationExisted) {
+        if (isUpStationExisted && isDownStationExisted) {
+            throw new RuntimeException("이미 등록된 구간 입니다.");
+        }
+        if (!getStations().isEmpty() && !isUpStationExisted && !isDownStationExisted) {
+            throw new RuntimeException("등록할 수 없는 구간 입니다.");
+        }
+    }
+
+    private List<Station> getStationsInOrder() {
+        List<Station> stations = new ArrayList<>();
+        Station downStation = findUpStation();
+        stations.add(downStation);
+
+        while (downStation != null) {
+            Station finalDownStation = downStation;
+            Optional<Section> nextLineStation = sections.stream()
+                    .filter(it -> it.getUpStation() == finalDownStation)
+                    .findFirst();
+            if (!nextLineStation.isPresent()) {
+                break;
+            }
+            downStation = nextLineStation.get().getDownStation();
+            stations.add(downStation);
+        }
+
+        return stations;
+    }
+
+    private void updateDownStation(Station upStation, Station downStation, int distance) {
+        sections.stream()
+                .filter(it -> it.getDownStation() == downStation)
+                .findFirst()
+                .ifPresent(it -> it.updateDownStation(upStation, distance));
+    }
+
+    private void updateUpStation(Station upStation, Station downStation, int distance) {
+        sections.stream()
+                .filter(it -> it.getUpStation() == upStation)
+                .findFirst()
+                .ifPresent(it -> it.updateUpStation(downStation, distance));
+    }
+
+    private void createSection(Line line, Section upLineStation, Section downLineStation) {
+        Station newUpStation = downLineStation.getUpStation();
+        Station newDownStation = upLineStation.getDownStation();
+        int newDistance = upLineStation.getDistance() + downLineStation.getDistance();
+        sections.add(new Section(line, newUpStation, newDownStation, newDistance));
+    }
+
+    private Optional<Section> findDownStation(Long stationId) {
+        return sections.stream()
+                .filter(it -> it.equalsDownStation(stationId))
+                .findFirst();
+    }
+
+    private Optional<Section> findUpStation(Long stationId) {
+        return sections.stream()
+                .filter(it -> it.equalsUpStation(stationId))
+                .findFirst();
+    }
+}
